Installation de laravel breeze et implementer la logique d'authentification

+++ app/Http/Controllers/PostController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use App\Models\Post;
 use App\Http\Requests\StorePostRequest;
 use App\Http\Resources\PostResource;
@@ -14,8 +15,12 @@ class PostController extends Controller
      */
     public function index()
     {
-        return PostResource::collection(Post::with('author')->paginate(1));
-            }
+        $user = request()->user();
+        //Si on veut retourner les infos de author
+        //$post = $user->posts()->with("author")->paginate();
+        $post = $user->posts()->paginate();
+        return PostResource::collection($post);
+    }
 
     /**
      * Store a newly created resource in storage.
@@ -23,7 +28,8 @@ class PostController extends Controller
     public function store(StorePostRequest $request)
     {
         $data = $request->validated();
-        $data["author_id"] = 1;
+        $user = $request->user();
+        $data['author_id'] = $user->id;
         $post = Post::create($data);
         return response()->json(new PostResource($post), 201);
     }
@@ -33,7 +39,10 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
-        
+        $user = request()->user();
+
+        abort_if($user->id !== $post->author_id, 403, 'Accès non autorisé');
+
         return new PostResource($post);
     }
 
@@ -42,6 +51,7 @@ class PostController extends Controller
      */
     public function update(StorePostRequest $request, Post $post)
     {
+        abort_if(Auth::id() !== $post->author_id, 403, 'Accès non autorisé');
         $data = $request->validated();
         $post->update($data);
         return new PostResource($post);
@@ -52,6 +62,7 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        abort_if(Auth::id() !== $post->author_id, 403, 'Accès non autorisé');
         $post->delete();
         return response()->noContent();
     }
